Debugged mod_const.py

+++ pynicamdc/hgrid/mod_const.py
@@ -10,7 +10,7 @@ class Const:
     def __init__(self,single):
 
         #< undefined value (int2, float32, float64))   
-        self.CONST_UNDEF2   = np.int8(-32768)
+        self.CONST_UNDEF2   = np.int16(-32768)
         self.CONST_UNDEF4   = np.float32(-9.9999E30)
         self.CONST_UNDEF8   = np.float64(-9.9999E30)
